Refs #124. Moved paragraph testing to its own module
